+++ scripts/blacklist/list.js
@@ -21,7 +21,7 @@ module.exports = {
 				Embed.addFields([
 					{
 						name: result[i].serverName,
-						value: `${emoji.interaction.ID}: ${result[i].serverId}\nReason: ${result[i].reason}\n\n`,
+						value: `${emoji.normal.ID}: ${result[i].serverId}\nReason: ${result[i].reason}\n\n`,
 					},
 				]);
 			}
@@ -41,7 +41,7 @@ module.exports = {
 				Embed.addFields([
 					{
 						name: result[i].username,
-						value: `${emoji.interaction.ID}: ${result[i].userId}\nReason: ${result[i].reason}\n\n`,
+						value: `${emoji.normal.ID}: ${result[i].userId}\nReason: ${result[i].reason}\n\n`,
 					},
 				]);
 			}