+++ scripts/connected/server.js
@@ -1,5 +1,6 @@
 const { MessageEmbed, CommandInteraction } = require('discord.js');
 const { paginate } = require('../../utils');
+const { normal } = require('../../emoji.json');
 
 module.exports = {
 	/**
@@ -20,29 +21,27 @@ module.exports = {
 			const searchCursor = await connectedList.find();
 			const result = await searchCursor.toArray();
 
-			function generateEmbed(db) {
-				const embeds = [];
-				let k = 5;
+			if (result.length === 0) return interaction.reply(`No connected servers yet ${normal.bruhcat} `);
 
-				for (let i = 0; i < db.length; i += 5) {
-					const current = db.slice(i, k);
+			const embeds = [];
+			let k = 5;
 
-					let j = i;
-					k += 5;
+			for (let i = 0; i < result.length; i += 5) {
+				const current = result.slice(i, k);
 
-					const fields = current.map(value => { return { name: `${++j}. ${value.serverName}`, value: `ServerID: ${value.serverId}\nChannel: ${value.channelName} \`(${value.channelId})\`` }; });
+				let j = i;
+				k += 5;
 
-					const embed = new MessageEmbed()
-						.setDescription(`Showing the current connected servers: ${j}-${k >= result.length ? result.length : k} / **${result.length}**`)
-						.setColor(0x2F3136)
-						.setFields(fields);
-					embeds.push(embed);
-				}
-				return embeds;
+				const fields = current.map(value => { return { name: `${++j}. ${value.serverName}`, value: `ServerID: ${value.serverId}\nChannel: ${value.channelName} \`(${value.channelId})\`` }; });
+
+				const embed = new MessageEmbed()
+					.setDescription(`Showing the current connected servers: ${j}-${k >= result.length ? result.length : k} / **${result.length}**`)
+					.setColor(0x2F3136)
+					.setFields(fields);
+				embeds.push(embed);
 			}
-
-			paginate(interaction, generateEmbed(result));
-
+			return paginate(interaction, embeds);
 		}
+
 	},
 };